Make prototype compile with dev branch.
